requirements.txt
fastapi
uvicorn
sqlalchemy
pydanticpip
